adds patient to investigation reponse
